Reimplement DOMRect/SVGRect
