+++ lib/jsdom/living/geometry/DOMRect-impl.js
@@ -3,11 +3,6 @@ const DOMRectReadOnlyImpl = require("./DOMRectReadOnly-impl").implementation;
 const DOMRect = require("../generated/DOMRect");
 
 class DOMRectImpl extends DOMRectReadOnlyImpl {
-  // https://svgwg.org/svg2-draft/types.html#TermReserialize
-  _reserialize({ x, y, width, height }) {
-    this._reflectedElement.setAttributeNS(null, this._reflectedAttribute, `${x} ${y} ${width} ${height}`);
-  }
-
   static fromRect(globalObject, other) {
     return DOMRect.createImpl(globalObject, [other.x, other.y, other.width, other.height]);
   }
@@ -15,72 +10,28 @@ class DOMRectImpl extends DOMRectReadOnlyImpl {
   get x() {
     return super.x;
   }
-
   set x(newX) {
-    if (this._reflectedElement) {
-      const { y, width, height } = this;
-      this._reserialize({
-        x: newX,
-        y,
-        width,
-        height
-      });
-      return;
-    }
     this._x = newX;
   }
 
   get y() {
     return super.y;
   }
-
   set y(newY) {
-    if (this._reflectedElement) {
-      const { x, width, height } = this;
-      this._reserialize({
-        x,
-        y: newY,
-        width,
-        height
-      });
-      return;
-    }
     this._y = newY;
   }
 
   get width() {
     return super.width;
   }
-
   set width(newWidth) {
-    if (this._reflectedElement) {
-      const { x, y, height } = this;
-      this._reserialize({
-        x,
-        y,
-        width: newWidth,
-        height
-      });
-      return;
-    }
     this._width = newWidth;
   }
 
   get height() {
     return super.height;
   }
-
   set height(newHeight) {
-    if (this._reflectedElement) {
-      const { x, y, width } = this;
-      this._reserialize({
-        x,
-        y,
-        width,
-        height: newHeight
-      });
-      return;
-    }
     this._height = newHeight;
   }
 }
