+++ lib/jsdom/living/geometry/DOMRectReadOnly-impl.js
@@ -2,17 +2,12 @@
 const DOMRectReadOnly = require("../generated/DOMRectReadOnly");
 
 class DOMRectReadOnlyImpl {
-  constructor(globalObject, [x = 0, y = 0, width = 0, height = 0], privateData) {
+  constructor(globalObject, [x = 0, y = 0, width = 0, height = 0]) {
     this._globalObject = globalObject;
     this._x = x;
     this._y = y;
     this._width = width;
     this._height = height;
-
-    // Reflection mode for SVG's viewBox.
-    this._reflectedElement = privateData.reflectedElement;
-    this._reflectedAttribute = privateData.reflectedAttribute;
-    this._parser = privateData.parser;
   }
 
   static fromRect(globalObject, other) {
@@ -20,34 +15,18 @@ class DOMRectReadOnlyImpl {
   }
 
   get x() {
-    if (this._reflectedElement) {
-      const attr = this._reflectedElement.getAttributeNS(null, this._reflectedAttribute);
-      return this._parser(attr).x;
-    }
     return this._x;
   }
 
   get y() {
-    if (this._reflectedElement) {
-      const attr = this._reflectedElement.getAttributeNS(null, this._reflectedAttribute);
-      return this._parser(attr).y;
-    }
     return this._y;
   }
 
   get width() {
-    if (this._reflectedElement) {
-      const attr = this._reflectedElement.getAttributeNS(null, this._reflectedAttribute);
-      return this._parser(attr).width;
-    }
     return this._width;
   }
 
   get height() {
-    if (this._reflectedElement) {
-      const attr = this._reflectedElement.getAttributeNS(null, this._reflectedAttribute);
-      return this._parser(attr).height;
-    }
     return this._height;
   }
