+++ lib/jsdom/living/svg/SVGAnimatedRect-impl.js
@@ -1,6 +1,5 @@
 "use strict";
 
-const DOMRectReadOnly = require("../generated/DOMRectReadOnly");
 const DOMRect = require("../generated/DOMRect");
 
 // https://drafts.csswg.org/css-syntax/#number-token-diagram
@@ -9,6 +8,7 @@ const numberRe = /^[+-]?(?:\d*\.)?\d+(?:[eE][+-]?\d+)?/;
 // https://svgwg.org/svg2-draft/coords.html#ViewBoxAttribute
 function parseViewBox(str) {
   // Use a do-while "loop" as JavaScript `goto end`.
+  // eslint-disable-next-line no-unreachable-loop
   do {
     // When the attribute does not exist.
     if (typeof str !== "string") {
@@ -110,10 +110,11 @@ class SVGAnimatedRectImpl {
   }
 
   get animVal() {
-    return DOMRectReadOnly.createImpl(this._globalObject, [], {
+    return DOMRect.createImpl(this._globalObject, [], {
       reflectedElement: this._element,
       reflectedAttribute: this._attribute,
-      parser: parseViewBox
+      parser: parseViewBox,
+      readOnly: true
     });
   }
 }
